+++ src/value/parsing.rs
@@ -15,6 +15,10 @@ use crate::value::primitives::*;
 use crate::value::traits::*;
 use chrono::{DateTime, Local, Utc};
 
+// fn iso8601_to_dkvalue(is08601_struct: &iso8601::DateTime) {
+
+// }
+
 fn jsvalue_to_dkvalue(jsvalue: &serde_json::Value) -> Value {
     match jsvalue {
         serde_json::Value::Null => Value::Missing(Empty::Expected),