+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Version 0.2
 
+- **0.2.1**: Added DateTime primitives. Put `chrono` integration into a separate
+  feature.
 - **0.2.0**: Datakit `Value`s can now be parsed from literals following JSON
   syntax. For example, a string `"137"` can be parsed into a
   `Value::Number(Numeric::(137))`. Parsing of arrays and objects/dictionaries is
@@ -26,9 +28,7 @@
   defined in the schema. Alternatively, there must be a "minimal" mode, where
   the validation checks, if a table contains **at least** the columns defined in the
   schema.
-- **0.2**: Add DateTime primitives and potentially decouple from `chrono`. Put
-  `chrono` integration into a separate feature.
-- **0.2.1**: Implement custom `serde` Serializer/Deserializer for
+- **0.2.2**: Implement custom `serde` Serializer/Deserializer for
   `datakit::table`. DSV and JSON at least. For JSON there should be options to
   either serialize "raw", i.e. default serialization of Datakit values to JSON,
   or a "d3.js compatible" version, where a table is a list of objects with the